spotguides: introduce SCMFactory
